+++ cmd/root.go
@@ -2,9 +2,12 @@ package cmd
 
 import (
 	"fmt"
-	"github.com/spf13/cobra"
 	"os"
-	"github.com/spf13/viper"
+
+	"github.com/spf13/cobra"
+	// "github.com/spf13/viper"
+
+	"github.com/alajmo/mani/core/dao"
 )
 
 const (
@@ -14,6 +17,8 @@ const (
 )
 
 var (
+	config dao.Config
+	configErr error
 	configFile string
 	rootCmd    = &cobra.Command{
 		Use:   appName,
@@ -29,28 +34,24 @@ func Execute() {
 	}
 }
 
-func initConfig() {
-	viper.SetConfigFile(configFile)
-	viper.AutomaticEnv()
-}
-
 func init() {
 	cobra.OnInitialize(initConfig)
-
 	rootCmd.PersistentFlags().StringVar(&configFile, "config", "", "config file (by default it checks current and all parent directories for mani.yaml|yml)")
-	// rootCmd.PersistentFlags().StringVar(&cfgFile   , "config", "", "config file (default is $HOME/.cobra.yaml)")
-	viper.BindPFlag("config", rootCmd.PersistentFlags().Lookup("config"))
 
 	rootCmd.AddCommand(
 		versionCmd(),
-		initCmd(),
 		completionCmd(),
-		execCmd(&configFile),
-		runCmd(&configFile),
-		listCmd(&configFile),
-		describeCmd(&configFile),
-		syncCmd(&configFile),
-		infoCmd(&configFile),
-		editCmd(&configFile),
+		initCmd(),
+		execCmd(&config, &configErr),
+		runCmd(&config, &configErr),
+		listCmd(&config, &configErr),
+		describeCmd(&config, &configErr),
+		syncCmd(&config, &configErr),
+		infoCmd(&config),
+		editCmd(&config, &configErr),
 	)
+}
+
+func initConfig() {
+	config, configErr = dao.ReadConfig(configFile)
 }